Added Units to chart
Added settings to .scatter(): "xUnits" for the units on the X-axis and
"yUnits" for units on the Y-axis. These settings take an array and print
the array items evenly spaced along their corresponding axis, such that
the first is at the bottom-left corner, and the last is at the right or
top edge.